Add support for mappings in YamlFile.js
It includes basic mapping as it's done in other plugins - template, schema.
On top of it there is support for "comment_prefix" in schema file that allows
to specify substring which indicates context comments for translators.
If strings starts with different prefix - it will be ignored and not added to
the source string as a comment.